# public/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {},
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"error",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"mongoose.Types.ObjectId": {
				"type": "string"
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "typescript-6",
		"version": "1.0.0",
		"description": "Ico Client Panel apis",
		"license": {
			"name": "license"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/admin/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Get user login",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/admin/sellerRegistration": {
			"post": {
				"operationId": "SellerRegistration",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "signin seller",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isBlocked": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"phoneNumber": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"isBlocked",
									"address",
									"phoneNumber",
									"password",
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/admin/sellerLogin": {
			"post": {
				"operationId": "SellerLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Get seller login",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/admin/userRegistration": {
			"post": {
				"operationId": "UserRegistration",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "signin user",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isBlocked": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"phoneNumber": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"isBlocked",
									"address",
									"phoneNumber",
									"password",
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/admin/userLogin": {
			"post": {
				"operationId": "UserLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Get user login",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/admin/product": {
			"post": {
				"operationId": "Product",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Get product",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"size": {
										"type": "string"
									},
									"type": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"_for": {
										"type": "string"
									},
									"isAvailable": {
										"type": "string"
									},
									"offer": {
										"type": "string"
									},
									"owner": {
										"type": "string"
									},
									"product_image": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/update": {
			"post": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Get update",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"_id": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"type": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"_for": {
										"type": "string"
									},
									"size": {
										"type": "string"
									},
									"isAvailable": {
										"type": "string"
									},
									"offer": {
										"type": "string"
									},
									"owner": {
										"type": "string"
									},
									"product_image": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/productSearch": {
			"get": {
				"operationId": "ProductSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "All product Search",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/Search": {
			"get": {
				"operationId": "Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Search Product by Name",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/Delete_Product": {
			"get": {
				"operationId": "Delete_Product",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Delete Product",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/admin_Search": {
			"get": {
				"operationId": "Admin_Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Admin Search",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/add_to_cart": {
			"post": {
				"operationId": "Add_to_cart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"quantity": {
										"type": "string"
									},
									"product_id": {
										"type": "string"
									},
									"user_id": {
										"type": "string"
									}
								},
								"required": [
									"quantity",
									"product_id",
									"user_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/admin/User_Cart_Search": {
			"get": {
				"operationId": "User_Cart_Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "User Cart Search",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "user_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/lookup": {
			"get": {
				"operationId": "Lookup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "User Cart Search",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/admin/name_Search": {
			"get": {
				"operationId": "Name_Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Name Search",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/fillProduct": {
			"post": {
				"operationId": "FillProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isBlocked": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"phoneNumber": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									},
									"userId": {
										"$ref": "#/components/schemas/mongoose.Types.ObjectId"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/admin/genderSearch": {
			"get": {
				"operationId": "GenderSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Gender search",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "_for",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/priceSearch": {
			"get": {
				"operationId": "PriceSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/offerOnProduct": {
			"post": {
				"operationId": "OfferOnProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"endDate": {
										"type": "string"
									},
									"startDate": {
										"type": "string"
									},
									"afterDiscount": {
										"type": "string"
									},
									"discount": {
										"type": "string"
									},
									"productId": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/admin/transactions": {
			"post": {
				"operationId": "Transactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"buyerId": {
										"type": "string"
									},
									"productId": {
										"type": "string"
									}
								},
								"required": [
									"buyerId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}